Remove outliers using point-by-point weighed outlier removal by conditional extremum

An outlier removal method that considers each potential outlier individually, using conditional thresholds based on top and bottom percentiles and error margins. It is a more refined alternative to simple percentile-based removal.

Usage

condextr(data, cols = NULL, group = NULL, top = 0.995,
top.error = 0.1, top.magnitude = 0.2, bottom = 0.0025, bottom.error = 0.2,
bottom.magnitude = 0.4, interval = 10, by = "min", half = 30,
times = 10, date_col = NULL, cores = NULL, verbose = FALSE)

Arguments

data

A data frame, matrix, or numeric vector. If a vector, outlier removal is applied directly.

cols

Column indices or names of numeric variables.

group

Grouping column index or name. Outlier removal is performed within each group.

top, top.error, top.magnitude

Parameters for the top threshold.

bottom, bottom.error, bottom.magnitude

Parameters for the bottom threshold.

interval

Number of outlier marking iterations between observation deletions.

by

Time unit for observation deletion. An invalid unit string raises an error.

half

Half window size in minutes for observation deletion.

times

Number of observation deletions.

date_col

Time column index or name.

cores

Number of CPU cores passed to the C++ backend. NULL (default) lets the backend choose based on data size.

verbose

Logical; if TRUE, prints timing and deletion counts.

Details

The method marks values as outliers if they are the maximum or minimum and exceed a threshold that depends on a high percentile plus an error margin. After a specified number of marking steps, observation deletion is applied to remove rows with excessive missing values created by outlier removal.

Value

A data frame with outliers removed.

References

1. Example data is from https://smear.avaa.csc.fi/download. It includes particle number concentrations in SMEAR I Varrio forest.

Author

Chun-Sheng Liang <chun-shengliang@qq.com>

Examples

condextr(obsedele(data[1:500, c(1, 4, 17:19)], cols = 3:5, group = 2),
         cols = 3:5, group = 2)
#>                    date    monthyear     3.98     4.47     5.01
#> 1   2019-12-31 16:00:00 January 2020       NA       NA       NA
#> 2   2019-12-31 16:10:00 January 2020       NA       NA       NA
#> 3   2019-12-31 16:20:00 January 2020       NA       NA       NA
#> 4   2019-12-31 16:30:00 January 2020  42.9722  74.7785  24.8375
#> 5   2019-12-31 16:40:00 January 2020       NA   6.8651  44.7176
#> 6   2019-12-31 16:50:00 January 2020  61.2601       NA       NA
#> 7   2019-12-31 17:00:00 January 2020 115.8960       NA       NA
#> 8   2019-12-31 17:10:00 January 2020       NA       NA       NA
#> 9   2019-12-31 17:20:00 January 2020  61.2601       NA       NA
#> 10  2019-12-31 17:30:00 January 2020  60.9405   6.5912  42.9332
#> 11  2019-12-31 17:40:00 January 2020       NA       NA       NA
#> 12  2019-12-31 17:50:00 January 2020       NA       NA       NA
#> 13  2019-12-31 18:00:00 January 2020       NA  13.9498  90.8651
#> 14  2019-12-31 18:10:00 January 2020       NA       NA       NA
#> 15  2019-12-31 18:20:00 January 2020  61.2601       NA       NA
#> 16  2019-12-31 18:30:00 January 2020 130.9630 227.8960  75.6951
#> 17  2019-12-31 18:40:00 January 2020       NA       NA       NA
#> 18  2019-12-31 18:50:00 January 2020       NA       NA       NA
#> 19  2019-12-31 19:00:00 January 2020       NA       NA       NA
#> 20  2019-12-31 19:10:00 January 2020 133.5230 148.2000  49.2241
#> 21  2019-12-31 19:20:00 January 2020  61.2601       NA       NA
#> 22  2019-12-31 19:40:00 January 2020       NA       NA       NA
#> 23  2019-12-31 19:50:00 January 2020       NA       NA       NA
#> 24  2019-12-31 20:00:00 January 2020       NA       NA       NA
#> 25  2019-12-31 20:10:00 January 2020       NA       NA       NA
#> 26  2019-12-31 20:20:00 January 2020  42.8723  74.6047  24.7797
#> 27  2019-12-31 20:30:00 January 2020  42.9722  74.7785  24.8375
#> 28  2019-12-31 20:40:00 January 2020       NA       NA       NA
#> 29  2019-12-31 20:50:00 January 2020       NA   7.2200  47.0292
#> 30  2019-12-31 21:00:00 January 2020       NA       NA       NA
#> 31  2019-12-31 21:10:00 January 2020  96.5514  72.1565  23.9666
#> 32  2019-12-31 21:20:00 January 2020  39.7904  75.6110  64.4859
#> 33  2019-12-31 21:30:00 January 2020  96.7000  72.5003  24.0808
#> 34  2019-12-31 21:40:00 January 2020       NA       NA       NA
#> 35  2019-12-31 21:50:00 January 2020       NA       NA       NA
#> 36  2019-12-31 22:00:00 January 2020       NA       NA       NA
#> 37  2019-12-31 22:10:00 January 2020       NA       NA       NA
#> 38  2019-12-31 22:20:00 January 2020  45.0185  78.3394  26.0202
#> 39  2019-12-31 22:30:00 January 2020  61.2601       NA       NA
#> 40  2019-12-31 22:40:00 January 2020  82.6597 150.5060  91.1891
#> 41  2019-12-31 22:50:00 January 2020       NA       NA       NA
#> 42  2019-12-31 23:00:00 January 2020  61.2601       NA       NA
#> 43  2019-12-31 23:10:00 January 2020       NA       NA       NA
#> 44  2020-01-01 00:10:00 January 2020       NA       NA       NA
#> 45  2020-01-01 00:20:00 January 2020       NA       NA       NA
#> 46  2020-01-01 00:30:00 January 2020       NA       NA       NA
#> 47  2020-01-01 00:40:00 January 2020  35.9966  76.5024 111.1020
#> 48  2020-01-01 00:50:00 January 2020       NA       NA       NA
#> 49  2020-01-01 01:00:00 January 2020       NA       NA       NA
#> 50  2020-01-01 01:10:00 January 2020  61.2601       NA       NA
#> 51  2020-01-01 01:20:00 January 2020  94.8684  76.2325  25.3204
#> 52  2020-01-01 01:30:00 January 2020       NA   6.6070  43.0364
#> 53  2020-01-01 01:40:00 January 2020       NA   6.9749  45.4325
#> 54  2020-01-01 01:50:00 January 2020       NA       NA       NA
#> 55  2020-01-01 03:50:00 January 2020       NA       NA       NA
#> 56  2020-01-01 04:00:00 January 2020       NA       NA       NA
#> 57  2020-01-01 04:10:00 January 2020       NA       NA       NA
#> 58  2020-01-01 04:20:00 January 2020 129.7860 148.3430  49.2718
#> 59  2020-01-01 04:30:00 January 2020       NA       NA       NA
#> 60  2020-01-01 04:40:00 January 2020       NA       NA       NA
#> 61  2020-01-01 04:50:00 January 2020       NA       NA       NA
#> 62  2020-01-01 06:50:00 January 2020       NA  13.9498  90.8651
#> 63  2020-01-01 07:00:00 January 2020       NA       NA       NA
#> 64  2020-01-01 07:10:00 January 2020       NA       NA       NA
#> 65  2020-01-01 07:20:00 January 2020 122.5200       NA       NA
#> 66  2020-01-01 07:30:00 January 2020       NA   7.2332  47.1152
#> 67  2020-01-01 07:40:00 January 2020  42.9722  74.7785  24.8375
#> 68  2020-01-01 07:50:00 January 2020       NA       NA       NA
#> 69  2020-01-01 08:00:00 January 2020       NA       NA       NA
#> 70  2020-01-01 08:10:00 January 2020       NA       NA       NA
#> 71  2020-01-01 10:20:00 January 2020       NA       NA       NA
#> 72  2020-01-01 10:30:00 January 2020       NA       NA       NA
#> 73  2020-01-01 10:40:00 January 2020       NA       NA       NA
#> 74  2020-01-01 10:50:00 January 2020  45.0185  78.3394  26.0202
#> 75  2020-01-01 11:00:00 January 2020       NA       NA       NA
#> 76  2020-01-01 11:10:00 January 2020       NA       NA       NA
#> 77  2020-01-01 11:20:00 January 2020       NA       NA       NA
#> 78  2020-01-01 11:40:00 January 2020       NA       NA       NA
#> 79  2020-01-01 11:50:00 January 2020       NA       NA       NA
#> 80  2020-01-01 12:00:00 January 2020       NA       NA       NA
#> 81  2020-01-01 12:10:00 January 2020  42.4098  73.7999  24.5124
#> 82  2020-01-01 12:20:00 January 2020  45.0185  78.3394  26.0202
#> 83  2020-01-01 12:30:00 January 2020       NA       NA       NA
#> 84  2020-01-01 12:40:00 January 2020       NA   6.8972  44.9267
#> 85  2020-01-01 12:50:00 January 2020       NA       NA       NA
#> 86  2020-01-01 13:50:00 January 2020       NA       NA       NA
#> 87  2020-01-01 14:00:00 January 2020       NA       NA       NA
#> 88  2020-01-01 14:10:00 January 2020       NA       NA       NA
#> 89  2020-01-01 14:20:00 January 2020  60.9227   6.9600  45.3353
#> 90  2020-01-01 14:30:00 January 2020       NA       NA       NA
#> 91  2020-01-01 14:40:00 January 2020       NA   7.2332  47.1152
#> 92  2020-01-01 14:50:00 January 2020       NA       NA       NA
#> 93  2020-01-01 15:50:00 January 2020       NA       NA       NA
#> 94  2020-01-01 16:00:00 January 2020       NA       NA       NA
#> 95  2020-01-01 16:10:00 January 2020  57.1449       NA       NA
#> 96  2020-01-01 16:20:00 January 2020       NA   9.3776  47.8526
#> 97  2020-01-01 16:30:00 January 2020       NA       NA       NA
#> 98  2020-01-01 16:40:00 January 2020       NA       NA       NA
#> 99  2020-01-01 16:50:00 January 2020  46.2837  72.0913  21.9483
#> 100 2020-01-01 17:00:00 January 2020       NA       NA       NA
#> 101 2020-01-01 17:10:00 January 2020       NA       NA       NA
#> 102 2020-01-01 17:20:00 January 2020  42.4445  75.6723  68.9153
#> 103 2020-01-01 17:30:00 January 2020       NA       NA       NA
#> 104 2020-01-01 17:40:00 January 2020       NA       NA       NA
#> 105 2020-01-01 17:50:00 January 2020       NA       NA       NA
#> 106 2020-01-01 19:30:00 January 2020       NA       NA       NA
#> 107 2020-01-01 19:40:00 January 2020       NA       NA       NA
#> 108 2020-01-01 19:50:00 January 2020       NA       NA       NA
#> 109 2020-01-01 20:00:00 January 2020  96.1242  69.8581  21.2684
#> 110 2020-01-01 20:10:00 January 2020       NA       NA       NA
#> 111 2020-01-01 20:20:00 January 2020       NA       NA       NA
#> 112 2020-01-01 20:30:00 January 2020       NA       NA       NA
#> 113 2020-01-01 20:40:00 January 2020       NA       NA       NA
#> 114 2020-01-01 20:50:00 January 2020       NA   9.3871  47.9009
#> 115 2020-01-01 21:00:00 January 2020       NA       NA       NA
#> 116 2020-01-01 21:10:00 January 2020  57.1449       NA       NA
#> 117 2020-01-01 21:20:00 January 2020       NA       NA       NA
#> 118 2020-01-02 00:20:00 January 2020       NA       NA       NA
#> 119 2020-01-02 00:30:00 January 2020       NA       NA       NA
#> 120 2020-01-02 00:40:00 January 2020       NA       NA       NA
#> 121 2020-01-02 00:50:00 January 2020  46.2837  72.0913  21.9483
#> 122 2020-01-02 01:00:00 January 2020  49.8677  77.6738  23.6479
#> 123 2020-01-02 01:10:00 January 2020       NA       NA       NA
#> 124 2020-01-02 01:20:00 January 2020       NA       NA       NA
#> 125 2020-01-02 01:30:00 January 2020  58.8257       NA       NA
#> 126 2020-01-02 01:50:00 January 2020       NA       NA       NA
#> 127 2020-01-02 02:00:00 January 2020       NA       NA       NA
#> 128 2020-01-02 02:10:00 January 2020       NA       NA       NA
#> 129 2020-01-02 02:20:00 January 2020  47.4833  73.9599  22.5172
#> 130 2020-01-02 02:30:00 January 2020  94.9316 147.8650  45.0177
#> 131 2020-01-02 02:40:00 January 2020  47.4722  73.9426  22.5119
#> 132 2020-01-02 02:50:00 January 2020       NA   9.4445  48.1940
#> 133 2020-01-02 03:00:00 January 2020       NA   9.1519  46.7010
#> 134 2020-01-02 03:10:00 January 2020       NA   9.4968  48.4608
#> 135 2020-01-02 03:20:00 January 2020  40.0351  81.8495 118.4440
#> 136 2020-01-02 03:30:00 January 2020       NA   9.2470  47.1860
#> 137 2020-01-02 03:40:00 January 2020       NA   9.1802  46.8450
#> 138 2020-01-02 03:50:00 January 2020  44.1765  77.8195  66.9276
#> 139 2020-01-02 04:00:00 January 2020  47.6785  74.2639  22.6097
#> 140 2020-01-02 04:10:00 January 2020       NA       NA       NA
#> 141 2020-01-02 04:20:00 January 2020  88.2793 154.9610 130.9460
#> 142 2020-01-02 04:30:00 January 2020  88.4717 152.8910 118.9460
#> 143 2020-01-02 04:40:00 January 2020 140.8420  86.1568 248.7880
#> 144 2020-01-02 04:50:00 January 2020  58.2736  24.1681 123.3260
#> 145 2020-01-02 05:00:00 January 2020  55.6383 103.1850       NA
#> 146 2020-01-02 05:10:00 January 2020       NA  14.9306  76.1888
#> 147 2020-01-02 05:20:00 January 2020 176.4080 236.2510 197.4570
#> 148 2020-01-02 05:30:00 January 2020  93.7203  78.1097 102.0320
#> 149 2020-01-02 05:40:00 January 2020       NA  25.0583 127.8690
#> 150 2020-01-02 05:50:00 January 2020  41.3755  80.5413 101.7510
#> 151 2020-01-02 06:00:00 January 2020       NA  35.4367 180.8280
#> 152 2020-01-02 06:10:00 January 2020       NA  14.2710  72.8230
#> 153 2020-01-02 06:20:00 January 2020 116.7100  28.6043 145.9640
#> 154 2020-01-02 06:30:00 January 2020       NA  43.7553 223.2770
#> 155 2020-01-02 06:40:00 January 2020 182.1380 227.2330  97.7361
#> 156 2020-01-02 06:50:00 January 2020 125.5730 163.8870 264.9040
#> 157 2020-01-02 07:00:00 January 2020  79.2616  93.9808 308.6480
#> 158 2020-01-02 07:10:00 January 2020  40.7583  81.0403 108.9100
#> 159 2020-01-02 07:20:00 January 2020  33.2681  88.3891 202.3910
#> 160 2020-01-02 07:30:00 January 2020       NA       NA       NA
#> 161 2020-01-02 07:40:00 January 2020       NA       NA       NA
#> 162 2020-01-02 07:50:00 January 2020       NA   9.1804  46.8460
#> 163 2020-01-02 08:00:00 January 2020       NA       NA       NA
#> 164 2020-01-02 08:10:00 January 2020       NA       NA       NA
#> 165 2020-01-02 08:20:00 January 2020  43.9970  77.3754  66.0025
#> 166 2020-01-02 08:30:00 January 2020       NA   9.0166  46.0105
#> 167 2020-01-02 08:40:00 January 2020       NA       NA       NA
#> 168 2020-01-02 08:50:00 January 2020       NA       NA       NA
#> 169 2020-01-02 09:00:00 January 2020       NA       NA       NA
#> 170 2020-01-02 09:10:00 January 2020       NA  19.2452  98.2054
#> 171 2020-01-02 09:20:00 January 2020       NA       NA       NA
#> 172 2020-01-02 09:30:00 January 2020  55.5838       NA       NA
#> 173 2020-01-02 09:40:00 January 2020       NA       NA       NA
#> 174 2020-01-02 09:50:00 January 2020  59.1985       NA       NA
#> 175 2020-01-02 10:00:00 January 2020  47.3995  73.8293  22.4774
#> 176 2020-01-02 10:10:00 January 2020       NA       NA       NA
#> 177 2020-01-02 10:20:00 January 2020       NA       NA       NA
#> 178 2020-01-02 10:30:00 January 2020       NA       NA       NA
#> 179 2020-01-02 13:20:00 January 2020       NA       NA       NA
#> 180 2020-01-02 13:30:00 January 2020       NA       NA       NA
#> 181 2020-01-02 13:40:00 January 2020       NA  20.0681 102.4050
#> 182 2020-01-02 13:50:00 January 2020  59.5532       NA       NA
#> 183 2020-01-02 14:00:00 January 2020       NA       NA       NA
#> 184 2020-01-02 14:10:00 January 2020       NA  10.0338  51.2011
#> 185 2020-01-02 14:20:00 January 2020       NA       NA       NA
#> 186 2020-01-02 15:10:00 January 2020       NA       NA       NA
#> 187 2020-01-02 15:20:00 January 2020       NA       NA       NA
#> 188 2020-01-02 15:30:00 January 2020       NA  10.0328  51.1961
#> 189 2020-01-02 15:40:00 January 2020  47.5941  74.1324  22.5697
#> 190 2020-01-02 15:50:00 January 2020       NA       NA       NA
#> 191 2020-01-02 16:00:00 January 2020       NA       NA       NA
#> 192 2020-01-02 16:10:00 January 2020  97.9560 157.3760  49.1198
#> 193 2020-01-02 16:20:00 January 2020       NA  18.4201 100.1020
#> 194 2020-01-02 16:30:00 January 2020       NA       NA       NA
#> 195 2020-01-02 16:40:00 January 2020       NA       NA       NA
#> 196 2020-01-02 17:30:00 January 2020       NA   9.0296  49.0703
#> 197 2020-01-02 17:40:00 January 2020       NA       NA       NA
#> 198 2020-01-02 17:50:00 January 2020       NA       NA       NA
#> 199 2020-01-02 18:00:00 January 2020  53.3735   9.0840  49.3660
#> 200 2020-01-02 18:10:00 January 2020  47.8390  76.8578  23.9887
#> 201 2020-01-02 18:20:00 January 2020       NA       NA       NA
#> 202 2020-01-02 18:30:00 January 2020       NA       NA       NA
#> 203 2020-01-02 18:40:00 January 2020       NA       NA       NA
#> 204 2020-01-02 18:50:00 January 2020       NA  18.8752 102.5750
#> 205 2020-01-02 19:00:00 January 2020       NA       NA       NA
#> 206 2020-01-02 19:10:00 January 2020       NA       NA       NA
#> 207 2020-01-02 19:20:00 January 2020  62.1791       NA       NA
#> 208 2020-01-02 19:30:00 January 2020       NA       NA       NA
#> 209 2020-01-02 19:40:00 January 2020 124.3580       NA       NA
#> 210 2020-01-02 19:50:00 January 2020       NA       NA       NA
#> 211 2020-01-02 20:00:00 January 2020  47.8390  76.8578  23.9887
#> 212 2020-01-02 20:10:00 January 2020       NA       NA       NA
#> 213 2020-01-02 20:20:00 January 2020       NA       NA       NA
#> 214 2020-01-02 20:30:00 January 2020       NA       NA       NA
#> 215 2020-01-02 20:40:00 January 2020       NA   9.0838  49.3650
#> 216 2020-01-02 20:50:00 January 2020  47.6323  76.5258  23.8851
#> 217 2020-01-02 21:00:00 January 2020  47.7325  76.6867  23.9353
#> 218 2020-01-02 21:10:00 January 2020       NA       NA       NA
#> 219 2020-01-02 21:20:00 January 2020  47.6365  76.5325  23.8872
#> 220 2020-01-02 21:30:00 January 2020       NA   9.0760  49.3225
#> 221 2020-01-02 21:40:00 January 2020  47.0453  75.5827  23.5907
#> 222 2020-01-02 21:50:00 January 2020       NA       NA       NA
#> 223 2020-01-02 22:00:00 January 2020       NA       NA       NA
#> 224 2020-01-02 22:10:00 January 2020       NA       NA       NA
#> 225 2020-01-03 00:40:00 January 2020       NA       NA       NA
#> 226 2020-01-03 00:50:00 January 2020       NA       NA       NA
#> 227 2020-01-03 01:00:00 January 2020       NA       NA       NA
#> 228 2020-01-03 01:10:00 January 2020  61.8423   9.0644  49.2592
#> 229 2020-01-03 01:20:00 January 2020       NA       NA       NA
#> 230 2020-01-03 01:30:00 January 2020       NA       NA       NA
#> 231 2020-01-03 01:40:00 January 2020       NA       NA       NA
#> 232 2020-01-03 01:50:00 January 2020       NA       NA       NA
#> 233 2020-01-03 02:00:00 January 2020       NA       NA       NA
#> 234 2020-01-03 02:10:00 January 2020  47.8390  76.8578  23.9887
#> 235 2020-01-03 02:20:00 January 2020  47.8390  76.8578  23.9887
#> 236 2020-01-03 02:30:00 January 2020       NA       NA       NA
#> 237 2020-01-03 02:40:00 January 2020       NA       NA       NA
#> 238 2020-01-03 02:50:00 January 2020       NA       NA       NA
#> 239 2020-01-03 03:00:00 January 2020       NA       NA       NA
#> 240 2020-01-03 03:10:00 January 2020  62.1791       NA       NA
#> 241 2020-01-03 03:20:00 January 2020       NA       NA       NA
#> 242 2020-01-03 03:30:00 January 2020  47.8390  76.8578  23.9887
#> 243 2020-01-03 03:40:00 January 2020       NA       NA       NA
#> 244 2020-01-03 03:50:00 January 2020       NA       NA       NA
#> 245 2020-01-03 04:00:00 January 2020       NA       NA       NA
#> 246 2020-01-03 04:10:00 January 2020  47.8390  76.8578  23.9887
#> 247 2020-01-03 04:20:00 January 2020       NA       NA       NA
#> 248 2020-01-03 04:30:00 January 2020       NA       NA       NA
#> 249 2020-01-03 04:40:00 January 2020       NA       NA       NA
#> 250 2020-01-03 06:30:00 January 2020  62.1791       NA       NA
#> 251 2020-01-03 06:40:00 January 2020       NA       NA       NA
#> 252 2020-01-03 06:50:00 January 2020  62.1791       NA       NA
#> 253 2020-01-03 07:00:00 January 2020       NA   9.4376  51.2876
#> 254 2020-01-03 07:10:00 January 2020       NA       NA       NA
#> 255 2020-01-03 07:20:00 January 2020       NA       NA       NA
#> 256 2020-01-03 07:30:00 January 2020  95.2210 152.9820  47.7483
#> 257 2020-01-03 07:40:00 January 2020       NA       NA       NA
#> 258 2020-01-03 07:50:00 January 2020       NA       NA       NA
#> 259 2020-01-03 08:00:00 January 2020       NA       NA       NA
#> 260 2020-01-03 11:30:00 January 2020       NA       NA       NA
#> 261 2020-01-03 11:40:00 January 2020       NA   8.5294  46.3523
#> 262 2020-01-03 11:50:00 January 2020       NA       NA       NA
#> 263 2020-01-03 12:00:00 January 2020 195.3800 313.8970  97.9729
#> 264 2020-01-03 12:10:00 January 2020  82.7348 160.5830 191.8090
#> 265 2020-01-03 12:20:00 January 2020  91.1554 154.2140  87.9038
#> 266 2020-01-03 12:30:00 January 2020 122.9290       NA       NA
#> 267 2020-01-03 12:40:00 January 2020 137.2390 156.5170  88.3913
#> 268 2020-01-03 12:50:00 January 2020       NA  24.1260 131.1100
#> 269 2020-01-03 13:00:00 January 2020  61.3205       NA       NA
#> 270 2020-01-03 13:10:00 January 2020  57.3812  15.5336  84.4157
#> 271 2020-01-03 13:20:00 January 2020       NA       NA       NA
#> 272 2020-01-03 13:30:00 January 2020  20.6152  95.4991 349.3280
#> 273 2020-01-03 13:40:00 January 2020       NA       NA       NA
#> 274 2020-01-03 13:50:00 January 2020       NA  18.2062  98.9394
#> 275 2020-01-03 14:00:00 January 2020       NA  16.6564  90.5174
#> 276 2020-01-03 14:10:00 January 2020       NA   7.5134  40.8306
#> 277 2020-01-03 14:20:00 January 2020       NA  53.7186 291.9280
#> 278 2020-01-03 14:30:00 January 2020 127.5410 247.7850 296.9720
#> 279 2020-01-03 14:40:00 January 2020  35.9767  83.4849 157.6220
#> 280 2020-01-03 14:50:00 January 2020 138.8220 151.4990  79.3879
#> 281 2020-01-03 15:00:00 January 2020       NA  15.5673  84.5990
#> 282 2020-01-03 15:10:00 January 2020  28.9170  88.1656 241.1560
#> 283 2020-01-03 15:20:00 January 2020  87.1124 155.2310 126.7040
#> 284 2020-01-03 15:30:00 January 2020  84.2456 157.4240 162.2140
#> 285 2020-01-03 15:40:00 January 2020 135.8130 153.5010 117.8450
#> 286 2020-01-03 15:50:00 January 2020  32.4445  86.0734 200.7580
#> 287 2020-01-03 16:00:00 January 2020  89.6049 150.6200  48.8608
#> 288 2020-01-03 16:10:00 January 2020  43.6422  73.3599  23.7977
#> 289 2020-01-03 16:20:00 January 2020       NA  39.0014 234.9950
#> 290 2020-01-03 16:30:00 January 2020       NA   5.3624  32.3102
#> 291 2020-01-03 16:40:00 January 2020 184.4770   5.0122  30.2000
#> 292 2020-01-03 16:50:00 January 2020  61.0394  12.4334  74.9147
#> 293 2020-01-03 17:00:00 January 2020       NA       NA       NA
#> 294 2020-01-03 17:10:00 January 2020  30.8797  81.6319 195.9400
#> 295 2020-01-03 17:20:00 January 2020  35.4143  80.0521 142.9670
#> 296 2020-01-03 17:30:00 January 2020  83.6945 153.6060 123.4890
#> 297 2020-01-03 17:40:00 January 2020 135.9390 150.8400  77.2636
#> 298 2020-01-03 17:50:00 January 2020  79.3617 156.1230 180.1740
#> 299 2020-01-03 18:00:00 January 2020 210.0160  71.7368 154.0070
#> 300 2020-01-03 18:10:00 January 2020  34.2377  80.6021 157.5560
#> 301 2020-01-03 18:20:00 January 2020 181.5690 305.2070  99.0082
#> 302 2020-01-03 18:30:00 January 2020       NA  13.9775  84.2183
#> 303 2020-01-03 18:40:00 January 2020       NA   4.9344  29.7313
#> 304 2020-01-03 18:50:00 January 2020 134.0560 231.7760 111.8770
#> 305 2020-01-03 19:00:00 January 2020  82.8914 154.4750 136.4230
#> 306 2020-01-03 19:10:00 January 2020  44.7403  75.2058  24.3965
#> 307 2020-01-03 19:20:00 January 2020  61.5626       NA       NA
#> 308 2020-01-03 19:30:00 January 2020 122.1780  22.5110 135.6360
#> 309 2020-01-03 19:40:00 January 2020       NA  36.3008 218.7230
#> 310 2020-01-03 19:50:00 January 2020  61.5626       NA       NA
#> 311 2020-01-03 20:00:00 January 2020       NA  13.5349  81.5515
#> 312 2020-01-03 20:10:00 January 2020 132.5740 152.5770 125.4130
#> 313 2020-01-03 20:20:00 January 2020 122.5210  14.3693  86.5795
#> 314 2020-01-03 20:30:00 January 2020  60.9670  14.1544  85.2842
#> 315 2020-01-03 20:40:00 January 2020       NA   5.7539  34.6690
#> 316 2020-01-03 20:50:00 January 2020  34.9161  80.2852 149.1460
#> 317 2020-01-03 21:00:00 January 2020 122.8610   6.2787  37.8309
#> 318 2020-01-03 21:10:00 January 2020 119.2200   6.7360  40.5865
#> 319 2020-01-03 21:20:00 January 2020  89.8514 151.0350  48.9952
#> 320 2020-01-03 21:30:00 January 2020  68.4794  89.8141 371.2460
#> 321 2020-01-03 21:40:00 January 2020  61.2524   7.3709  44.4120
#> 322 2020-01-03 21:50:00 January 2020  96.1658  74.7936  64.7557
#> 323 2020-01-03 22:00:00 January 2020  79.4507 156.0810 179.0700
#> 324 2020-01-03 22:10:00 January 2020 221.9960       NA 241.1480
#> 325 2020-01-03 22:20:00 January 2020 136.3510 229.1980  74.3512
#> 326 2020-01-03 22:30:00 January 2020       NA  22.1974 133.7460
#> 327 2020-01-03 22:40:00 January 2020  60.9297  15.0388  90.6133
#> 328 2020-01-03 22:50:00 January 2020  93.4064  74.4679  52.2496
#> 329 2020-01-03 23:00:00 January 2020       NA  14.7335  88.7736
#> 330 2020-01-03 23:10:00 January 2020       NA  15.4056  92.8232
#> 331 2020-01-03 23:20:00 January 2020  90.2795  77.8317 148.9960
#> 332 2020-01-03 23:30:00 January 2020  90.7859  75.9793  94.6833
#> 333 2020-01-03 23:40:00 January 2020       NA  30.4233 183.3090
#> 334 2020-01-03 23:50:00 January 2020  31.0337  82.0975 197.2700
#> 335 2020-01-04 00:00:00 January 2020  41.7805  77.0812  64.0602
#> 336 2020-01-04 00:10:00 January 2020       NA       NA       NA
#> 337 2020-01-04 00:20:00 January 2020       NA   5.4769  32.9997
#> 338 2020-01-04 00:30:00 January 2020  85.9354  78.3376 159.6100
#> 339 2020-01-04 00:40:00 January 2020       NA   7.3872  44.5099
#> 340 2020-01-04 00:50:00 January 2020       NA   6.9702  41.9974
#> 341 2020-01-04 01:00:00 January 2020  86.0127  78.2977 158.5000
#> 342 2020-01-04 01:10:00 January 2020       NA  32.0207 192.9340
#> 343 2020-01-04 01:20:00 January 2020       NA       NA       NA
#> 344 2020-01-04 01:30:00 January 2020  41.7468  77.0971  64.4788
#> 345 2020-01-04 01:40:00 January 2020  89.3610 150.2100  48.7278
#> 346 2020-01-04 01:50:00 January 2020       NA       NA       NA
#> 347 2020-01-04 02:00:00 January 2020       NA       NA       NA
#> 348 2020-01-04 02:10:00 January 2020       NA  15.4276  92.9555
#> 349 2020-01-04 02:20:00 January 2020       NA   7.2882  43.9137
#> 350 2020-01-04 02:30:00 January 2020  44.5442  74.8761  24.2896
#> 351 2020-01-04 02:40:00 January 2020       NA   6.9885  42.1076
#> 352 2020-01-04 02:50:00 January 2020       NA       NA       NA
#> 353 2020-01-04 03:00:00 January 2020  42.4322  76.7767  55.9799
#> 354 2020-01-04 03:10:00 January 2020       NA       NA       NA
#> 355 2020-01-04 03:20:00 January 2020       NA       NA       NA
#> 356 2020-01-04 03:30:00 January 2020  44.8402  75.3736  24.4510
#> 357 2020-01-04 03:40:00 January 2020       NA   7.6687  46.2059
#> 358 2020-01-04 03:50:00 January 2020  89.8514 151.0350  48.9952
#> 359 2020-01-04 04:00:00 January 2020       NA   8.0249  48.3525
#> 360 2020-01-04 04:10:00 January 2020       NA   7.4369  44.8093