Simple time series decomposition

Decomposes a time series into trend, seasonal, and residual components using simple moving averages and group means. The function returns the residual component (or the original series with trend removed) to simplify analysis.

Usage

decompose_ts(data, cols = NULL, date_col = NULL,
             period = "month", method = "additive", verbose = FALSE)

Arguments

data

A data frame with a time column.

cols

Column indices or names to decompose.

date_col

Time column index or name.

period

Seasonal period: "month", "day", or "hour".

method

Decomposition method: "additive" or "multiplicative".

verbose

Logical; if TRUE, prints progress message.

Value

A data frame with the residual component.

Examples

decompose_ts(data[1:500, c(1, 4, 17:19)], cols = 3:5, date_col = 1, period = "month")
#>                    date    monthyear         3.98         4.47        5.01
#> 1   2019-12-31 16:00:00 January 2020           NA           NA          NA
#> 2   2019-12-31 16:10:00 January 2020           NA           NA          NA
#> 3   2019-12-31 16:20:00 January 2020           NA           NA          NA
#> 4   2019-12-31 16:30:00 January 2020  -70.8730100  -77.4283571  -46.761543
#> 5   2019-12-31 16:40:00 January 2020           NA -111.3850571  -36.821493
#> 6   2019-12-31 16:50:00 January 2020  -61.7290600           NA          NA
#> 7   2019-12-31 17:00:00 January 2020  -28.3531100           NA          NA
#> 8   2019-12-31 17:10:00 January 2020           NA           NA          NA
#> 9   2019-12-31 17:20:00 January 2020  -79.9600100           NA          NA
#> 10  2019-12-31 17:30:00 January 2020  -78.3982900 -100.2487571  -41.324443
#> 11  2019-12-31 17:40:00 January 2020           NA           NA          NA
#> 12  2019-12-31 17:50:00 January 2020           NA           NA          NA
#> 13  2019-12-31 18:00:00 January 2020           NA  -89.0247071   -6.734793
#> 14  2019-12-31 18:10:00 January 2020           NA           NA          NA
#> 15  2019-12-31 18:20:00 January 2020  -76.8777433           NA          NA
#> 16  2019-12-31 18:30:00 January 2020  -21.8399767   86.6421179  -34.619193
#> 17  2019-12-31 18:40:00 January 2020           NA           NA          NA
#> 18  2019-12-31 18:50:00 January 2020           NA           NA          NA
#> 19  2019-12-31 19:00:00 January 2020           NA           NA          NA
#> 20  2019-12-31 19:10:00 January 2020  -26.9393500  -28.3876071  -62.216818
#> 21  2019-12-31 19:20:00 January 2020  -99.2022500           NA          NA
#> 22  2019-12-31 19:40:00 January 2020           NA           NA          NA
#> 23  2019-12-31 19:50:00 January 2020           NA           NA          NA
#> 24  2019-12-31 20:00:00 January 2020           NA           NA          NA
#> 25  2019-12-31 20:10:00 January 2020           NA           NA          NA
#> 26  2019-12-31 20:20:00 January 2020 -113.9764100 -153.0572238  -71.881476
#> 27  2019-12-31 20:30:00 January 2020 -110.2189300 -134.0196571  -65.558143
#> 28  2019-12-31 20:40:00 January 2020           NA           NA          NA
#> 29  2019-12-31 20:50:00 January 2020           NA -146.4091571  -36.199968
#> 30  2019-12-31 21:00:00 January 2020           NA           NA          NA
#> 31  2019-12-31 21:10:00 January 2020  -49.7574100  -80.6637971  -56.762363
#> 32  2019-12-31 21:20:00 January 2020  -87.7718900  -62.6914971  -19.295423
#> 33  2019-12-31 21:30:00 January 2020  -37.9502700  -67.7398905  -57.544026
#> 34  2019-12-31 21:40:00 January 2020           NA           NA          NA
#> 35  2019-12-31 21:50:00 January 2020           NA           NA          NA
#> 36  2019-12-31 22:00:00 January 2020           NA           NA          NA
#> 37  2019-12-31 22:10:00 January 2020           NA           NA          NA
#> 38  2019-12-31 22:20:00 January 2020  -90.0610100  -62.5232405  -55.811376
#> 39  2019-12-31 22:30:00 January 2020  -77.4769900           NA          NA
#> 40  2019-12-31 22:40:00 January 2020  -58.5433267   -2.9778905   -1.701076
#> 41  2019-12-31 22:50:00 January 2020           NA           NA          NA
#> 42  2019-12-31 23:00:00 January 2020  -78.6472243           NA          NA
#> 43  2019-12-31 23:10:00 January 2020           NA           NA          NA
#> 44  2019-12-31 23:20:00 January 2020  -80.5391600           NA          NA
#> 45  2019-12-31 23:30:00 January 2020           NA           NA          NA
#> 46  2019-12-31 23:40:00 January 2020           NA           NA          NA
#> 47  2019-12-31 23:50:00 January 2020  -71.1527100           NA          NA
#> 48  2020-01-01 00:00:00 January 2020           NA           NA          NA
#> 49  2020-01-01 00:10:00 January 2020           NA           NA          NA
#> 50  2020-01-01 00:20:00 January 2020           NA           NA          NA
#> 51  2020-01-01 00:30:00 January 2020           NA           NA          NA
#> 52  2020-01-01 00:40:00 January 2020  -96.0852703  -67.4552178  -92.417615
#> 53  2020-01-01 00:50:00 January 2020           NA           NA          NA
#> 54  2020-01-01 01:00:00 January 2020           NA           NA          NA
#> 55  2020-01-01 01:10:00 January 2020  -70.8217703           NA          NA
#> 56  2020-01-01 01:20:00 January 2020  -46.4854953  -67.5901678 -135.308415
#> 57  2020-01-01 01:30:00 January 2020           NA -113.9621845 -109.200815
#> 58  2020-01-01 01:40:00 January 2020           NA -102.0595178 -103.207940
#> 59  2020-01-01 01:50:00 January 2020           NA           NA          NA
#> 60  2020-01-01 02:20:00 January 2020           NA           NA          NA
#> 61  2020-01-01 02:30:00 January 2020           NA           NA          NA
#> 62  2020-01-01 02:40:00 January 2020           NA           NA          NA
#> 63  2020-01-01 02:50:00 January 2020           NA           NA          NA
#> 64  2020-01-01 03:00:00 January 2020           NA           NA          NA
#> 65  2020-01-01 03:10:00 January 2020           NA           NA          NA
#> 66  2020-01-01 03:20:00 January 2020           NA           NA          NA
#> 67  2020-01-01 03:30:00 January 2020  -64.1817953           NA          NA
#> 68  2020-01-01 03:40:00 January 2020           NA           NA          NA
#> 69  2020-01-01 03:50:00 January 2020           NA           NA          NA
#> 70  2020-01-01 04:00:00 January 2020           NA           NA          NA
#> 71  2020-01-01 04:10:00 January 2020           NA           NA          NA
#> 72  2020-01-01 04:20:00 January 2020  -74.3745953  -67.4552178  -92.417615
#> 73  2020-01-01 04:30:00 January 2020           NA           NA          NA
#> 74  2020-01-01 04:40:00 January 2020           NA           NA          NA
#> 75  2020-01-01 04:50:00 January 2020           NA           NA          NA
#> 76  2020-01-01 05:00:00 January 2020           NA           NA          NA
#> 77  2020-01-01 05:40:00 January 2020           NA           NA          NA
#> 78  2020-01-01 05:50:00 January 2020           NA           NA          NA
#> 79  2020-01-01 06:00:00 January 2020           NA           NA          NA
#> 80  2020-01-01 06:10:00 January 2020           NA           NA          NA
#> 81  2020-01-01 06:20:00 January 2020           NA -138.1392678  -94.337265
#> 82  2020-01-01 06:30:00 January 2020           NA           NA          NA
#> 83  2020-01-01 06:40:00 January 2020           NA -105.1058511  -31.999048
#> 84  2020-01-01 06:50:00 January 2020           NA  -67.5413178  -92.978382
#> 85  2020-01-01 07:00:00 January 2020           NA           NA          NA
#> 86  2020-01-01 07:10:00 January 2020           NA           NA          NA
#> 87  2020-01-01 07:20:00 January 2020  -78.0075953           NA          NA
#> 88  2020-01-01 07:30:00 January 2020           NA  -72.5572428 -125.650615
#> 89  2020-01-01 07:40:00 January 2020 -117.7814953  -17.5005978 -136.826175
#> 90  2020-01-01 07:50:00 January 2020           NA           NA          NA
#> 91  2020-01-01 08:00:00 January 2020           NA           NA          NA
#> 92  2020-01-01 08:10:00 January 2020           NA           NA          NA
#> 93  2020-01-01 08:20:00 January 2020           NA           NA          NA
#> 94  2020-01-01 08:30:00 January 2020           NA           NA          NA
#> 95  2020-01-01 08:40:00 January 2020           NA           NA          NA
#> 96  2020-01-01 08:50:00 January 2020           NA           NA          NA
#> 97  2020-01-01 09:00:00 January 2020           NA           NA          NA
#> 98  2020-01-01 09:10:00 January 2020           NA           NA          NA
#> 99  2020-01-01 09:20:00 January 2020           NA           NA          NA
#> 100 2020-01-01 09:30:00 January 2020           NA           NA          NA
#> 101 2020-01-01 09:40:00 January 2020           NA           NA          NA
#> 102 2020-01-01 09:50:00 January 2020           NA           NA          NA
#> 103 2020-01-01 10:00:00 January 2020           NA           NA          NA
#> 104 2020-01-01 10:10:00 January 2020           NA           NA          NA
#> 105 2020-01-01 10:20:00 January 2020           NA           NA          NA
#> 106 2020-01-01 10:30:00 January 2020           NA           NA          NA
#> 107 2020-01-01 10:40:00 January 2020           NA           NA          NA
#> 108 2020-01-01 10:50:00 January 2020  -78.0075953  -67.4552178  -92.417615
#> 109 2020-01-01 11:00:00 January 2020           NA           NA          NA
#> 110 2020-01-01 11:10:00 January 2020           NA           NA          NA
#> 111 2020-01-01 11:20:00 January 2020           NA           NA          NA
#> 112 2020-01-01 11:30:00 January 2020           NA           NA          NA
#> 113 2020-01-01 11:40:00 January 2020           NA           NA          NA
#> 114 2020-01-01 11:50:00 January 2020           NA           NA          NA
#> 115 2020-01-01 12:00:00 January 2020           NA           NA          NA
#> 116 2020-01-01 12:10:00 January 2020  -79.3119453  -69.7249678  -93.171515
#> 117 2020-01-01 12:20:00 January 2020  -77.1380287  -65.9420511  -91.915015
#> 118 2020-01-01 12:30:00 January 2020           NA           NA          NA
#> 119 2020-01-01 12:40:00 January 2020           NA -119.9019928  -77.860790
#> 120 2020-01-01 12:50:00 January 2020           NA           NA          NA
#> 121 2020-01-01 13:00:00 January 2020           NA           NA          NA
#> 122 2020-01-01 13:10:00 January 2020           NA           NA          NA
#> 123 2020-01-01 13:20:00 January 2020           NA           NA          NA
#> 124 2020-01-01 13:30:00 January 2020           NA           NA          NA
#> 125 2020-01-01 13:40:00 January 2020           NA           NA          NA
#> 126 2020-01-01 13:50:00 January 2020           NA           NA          NA
#> 127 2020-01-01 14:00:00 January 2020           NA           NA          NA
#> 128 2020-01-01 14:10:00 January 2020           NA           NA          NA
#> 129 2020-01-01 14:20:00 January 2020  -78.0075953  -67.4238178  -92.213315
#> 130 2020-01-01 14:30:00 January 2020           NA           NA          NA
#> 131 2020-01-01 14:40:00 January 2020           NA  -67.3186178  -91.527665
#> 132 2020-01-01 14:50:00 January 2020           NA           NA          NA
#> 133 2020-01-01 15:00:00 January 2020           NA           NA          NA
#> 134 2020-01-01 15:10:00 January 2020           NA           NA          NA
#> 135 2020-01-01 15:20:00 January 2020           NA           NA          NA
#> 136 2020-01-01 15:30:00 January 2020           NA  -67.5581511  -93.087915
#> 137 2020-01-01 15:40:00 January 2020           NA           NA          NA
#> 138 2020-01-01 15:50:00 January 2020           NA           NA          NA
#> 139 2020-01-01 16:00:00 January 2020           NA           NA          NA
#> 140 2020-01-01 16:10:00 January 2020  -79.8964953           NA          NA
#> 141 2020-01-01 16:20:00 January 2020           NA  -65.9286178  -91.294215
#> 142 2020-01-01 16:30:00 January 2020           NA           NA          NA
#> 143 2020-01-01 16:40:00 January 2020           NA           NA          NA
#> 144 2020-01-01 16:50:00 January 2020  -83.4381953  -24.8342845 -108.809548
#> 145 2020-01-01 17:00:00 January 2020           NA           NA          NA
#> 146 2020-01-01 17:10:00 January 2020           NA           NA          NA
#> 147 2020-01-01 17:20:00 January 2020  -84.1874620  -32.8037678  -69.486315
#> 148 2020-01-01 17:30:00 January 2020           NA           NA          NA
#> 149 2020-01-01 17:40:00 January 2020           NA           NA          NA
#> 150 2020-01-01 17:50:00 January 2020           NA           NA          NA
#> 151 2020-01-01 18:00:00 January 2020           NA           NA          NA
#> 152 2020-01-01 18:10:00 January 2020           NA           NA          NA
#> 153 2020-01-01 18:20:00 January 2020  -69.4870620           NA          NA
#> 154 2020-01-01 18:30:00 January 2020           NA           NA          NA
#> 155 2020-01-01 18:40:00 January 2020           NA           NA          NA
#> 156 2020-01-01 18:50:00 January 2020           NA           NA          NA
#> 157 2020-01-01 19:00:00 January 2020           NA           NA          NA
#> 158 2020-01-01 19:10:00 January 2020           NA           NA          NA
#> 159 2020-01-01 19:20:00 January 2020           NA           NA          NA
#> 160 2020-01-01 19:30:00 January 2020           NA           NA          NA
#> 161 2020-01-01 19:40:00 January 2020           NA           NA          NA
#> 162 2020-01-01 19:50:00 January 2020           NA           NA          NA
#> 163 2020-01-01 20:00:00 January 2020  -58.5179453  -67.4552178  -92.417615
#> 164 2020-01-01 20:10:00 January 2020           NA           NA          NA
#> 165 2020-01-01 20:20:00 January 2020           NA           NA          NA
#> 166 2020-01-01 20:30:00 January 2020           NA           NA          NA
#> 167 2020-01-01 20:40:00 January 2020           NA           NA          NA
#> 168 2020-01-01 20:50:00 January 2020           NA  -97.6907178  -79.101365
#> 169 2020-01-01 21:00:00 January 2020           NA           NA          NA
#> 170 2020-01-01 21:10:00 January 2020  -97.4972453           NA          NA
#> 171 2020-01-01 21:20:00 January 2020           NA           NA          NA
#> 172 2020-01-01 21:30:00 January 2020           NA           NA          NA
#> 173 2020-01-01 21:40:00 January 2020           NA           NA          NA
#> 174 2020-01-01 21:50:00 January 2020           NA           NA          NA
#> 175 2020-01-01 22:00:00 January 2020           NA  -67.4552178  -92.417615
#> 176 2020-01-01 22:10:00 January 2020           NA           NA          NA
#> 177 2020-01-01 22:20:00 January 2020           NA           NA          NA
#> 178 2020-01-01 22:30:00 January 2020           NA           NA          NA
#> 179 2020-01-01 22:40:00 January 2020           NA           NA          NA
#> 180 2020-01-01 22:50:00 January 2020           NA           NA          NA
#> 181 2020-01-01 23:00:00 January 2020           NA           NA          NA
#> 182 2020-01-01 23:10:00 January 2020           NA           NA          NA
#> 183 2020-01-01 23:20:00 January 2020           NA           NA          NA
#> 184 2020-01-01 23:30:00 January 2020           NA           NA          NA
#> 185 2020-01-01 23:40:00 January 2020           NA           NA          NA
#> 186 2020-01-01 23:50:00 January 2020           NA           NA          NA
#> 187 2020-01-02 00:00:00 January 2020  -78.0075953           NA          NA
#> 188 2020-01-02 00:10:00 January 2020           NA           NA          NA
#> 189 2020-01-02 00:20:00 January 2020           NA           NA          NA
#> 190 2020-01-02 00:30:00 January 2020           NA           NA          NA
#> 191 2020-01-02 00:40:00 January 2020           NA           NA          NA
#> 192 2020-01-02 00:50:00 January 2020  -83.4381953  -67.4552178  -92.417615
#> 193 2020-01-02 01:00:00 January 2020  -79.2386620  -64.6639678  -91.567815
#> 194 2020-01-02 01:10:00 January 2020           NA           NA          NA
#> 195 2020-01-02 01:20:00 January 2020           NA           NA          NA
#> 196 2020-01-02 01:30:00 January 2020  -72.2123953           NA          NA
#> 197 2020-01-02 01:40:00 January 2020           NA           NA          NA
#> 198 2020-01-02 01:50:00 January 2020           NA           NA          NA
#> 199 2020-01-02 02:00:00 January 2020           NA           NA          NA
#> 200 2020-01-02 02:10:00 January 2020           NA           NA          NA
#> 201 2020-01-02 02:20:00 January 2020  -81.1393953  -68.0703178  -92.604882
#> 202 2020-01-02 02:30:00 January 2020  -42.5543953  -12.4877178  -75.682690
#> 203 2020-01-02 02:40:00 January 2020  -88.0127620  -82.6191378  -97.034315
#> 204 2020-01-02 02:50:00 January 2020           NA -134.5878778  -76.601355
#> 205 2020-01-02 03:00:00 January 2020           NA -121.1760978  -82.704975
#> 206 2020-01-02 03:10:00 January 2020           NA -111.9352011  -82.857248
#> 207 2020-01-02 03:20:00 January 2020  -95.7220753  -43.5643178  -24.237415
#> 208 2020-01-02 03:30:00 January 2020           NA -110.0778678  -95.110690
#> 209 2020-01-02 03:40:00 January 2020           NA -105.4013956  -95.114570
#> 210 2020-01-02 03:50:00 January 2020  -88.6508353  -39.8314078  -76.770535
#> 211 2020-01-02 04:00:00 January 2020  -83.9586287  -45.5750269 -118.481997
#> 212 2020-01-02 04:10:00 January 2020           NA           NA          NA
#> 213 2020-01-02 04:20:00 January 2020  -50.1571620   27.7583367  -20.002860
#> 214 2020-01-02 04:30:00 January 2020  -48.8881120   25.2314276  -38.723615
#> 215 2020-01-02 04:40:00 January 2020  -12.0794453  -42.6131542   70.547831
#> 216 2020-01-02 04:50:00 January 2020  -92.2706668 -105.9403633  -61.744351
#> 217 2020-01-02 05:00:00 January 2020  -92.7936703  -35.4719269  297.844285
#> 218 2020-01-02 05:10:00 January 2020           NA -124.2203087 -155.023642
#> 219 2020-01-02 05:20:00 January 2020   10.9294172   83.0635913  -40.938442
#> 220 2020-01-02 05:30:00 January 2020  -72.4526509  -81.3379542 -141.349442
#> 221 2020-01-02 05:40:00 January 2020           NA -135.8328178 -122.878260
#> 222 2020-01-02 05:50:00 January 2020 -124.4862287  -80.5972542 -152.162024
#> 223 2020-01-02 06:00:00 January 2020           NA -122.1721087  -87.468506
#> 224 2020-01-02 06:10:00 January 2020           NA -137.0142595 -186.885515
#> 225 2020-01-02 06:20:00 January 2020  -57.7275203 -112.1512345 -114.996015
#> 226 2020-01-02 06:30:00 January 2020           NA  -87.9055928  -46.377265
#> 227 2020-01-02 06:40:00 January 2020    0.6641618   83.8157572 -159.330507
#> 228 2020-01-02 06:50:00 January 2020  -65.5150382    8.8265155   -3.960773
#> 229 2020-01-02 07:00:00 January 2020 -115.2011953  -60.3126678   57.940560
#> 230 2020-01-02 07:10:00 January 2020 -144.2423828  -78.7623095 -144.524207
#> 231 2020-01-02 07:20:00 January 2020 -133.8400953  -59.0916845  -51.454373
#> 232 2020-01-02 07:30:00 January 2020           NA           NA          NA
#> 233 2020-01-02 07:40:00 January 2020           NA           NA          NA
#> 234 2020-01-02 07:50:00 January 2020           NA -136.8526078 -210.804325
#> 235 2020-01-02 08:00:00 January 2020           NA           NA          NA
#> 236 2020-01-02 08:10:00 January 2020           NA           NA          NA
#> 237 2020-01-02 08:20:00 January 2020 -118.1765953  -88.1849803 -191.254440
#> 238 2020-01-02 08:30:00 January 2020           NA -152.2014428 -189.088127
#> 239 2020-01-02 08:40:00 January 2020           NA           NA          NA
#> 240 2020-01-02 08:50:00 January 2020           NA           NA          NA
#> 241 2020-01-02 09:00:00 January 2020           NA           NA          NA
#> 242 2020-01-02 09:10:00 January 2020           NA  -88.8513578  -86.103295
#> 243 2020-01-02 09:20:00 January 2020           NA           NA          NA
#> 244 2020-01-02 09:30:00 January 2020  -72.2141953           NA          NA
#> 245 2020-01-02 09:40:00 January 2020           NA           NA          NA
#> 246 2020-01-02 09:50:00 January 2020  -71.7355287           NA          NA
#> 247 2020-01-02 10:00:00 January 2020  -82.1527953  -38.4925428 -128.114165
#> 248 2020-01-02 10:10:00 January 2020           NA           NA          NA
#> 249 2020-01-02 10:20:00 January 2020           NA           NA          NA
#> 250 2020-01-02 10:30:00 January 2020           NA           NA          NA
#> 251 2020-01-02 10:40:00 January 2020           NA           NA          NA
#> 252 2020-01-02 10:50:00 January 2020  -29.5347953           NA          NA
#> 253 2020-01-02 11:00:00 January 2020           NA           NA          NA
#> 254 2020-01-02 11:10:00 January 2020           NA           NA          NA
#> 255 2020-01-02 11:20:00 January 2020           NA           NA          NA
#> 256 2020-01-02 11:30:00 January 2020           NA  -99.3795678  -78.191465
#> 257 2020-01-02 11:40:00 January 2020           NA           NA          NA
#> 258 2020-01-02 11:50:00 January 2020           NA           NA          NA
#> 259 2020-01-02 12:00:00 January 2020           NA           NA          NA
#> 260 2020-01-02 12:10:00 January 2020           NA           NA          NA
#> 261 2020-01-02 12:20:00 January 2020           NA           NA          NA
#> 262 2020-01-02 12:30:00 January 2020           NA           NA          NA
#> 263 2020-01-02 12:40:00 January 2020           NA           NA          NA
#> 264 2020-01-02 12:50:00 January 2020           NA           NA          NA
#> 265 2020-01-02 13:00:00 January 2020           NA           NA          NA
#> 266 2020-01-02 13:10:00 January 2020           NA           NA          NA
#> 267 2020-01-02 13:20:00 January 2020           NA           NA          NA
#> 268 2020-01-02 13:30:00 January 2020           NA           NA          NA
#> 269 2020-01-02 13:40:00 January 2020           NA  -67.4552178  -92.417615
#> 270 2020-01-02 13:50:00 January 2020  -78.0075953           NA          NA
#> 271 2020-01-02 14:00:00 January 2020           NA           NA          NA
#> 272 2020-01-02 14:10:00 January 2020           NA  -72.4723678 -118.019565
#> 273 2020-01-02 14:20:00 January 2020           NA           NA          NA
#> 274 2020-01-02 14:30:00 January 2020           NA  -71.0394511 -110.707782
#> 275 2020-01-02 14:40:00 January 2020           NA           NA          NA
#> 276 2020-01-02 14:50:00 January 2020           NA           NA          NA
#> 277 2020-01-02 15:00:00 January 2020           NA  -70.2410428 -106.633465
#> 278 2020-01-02 15:10:00 January 2020           NA           NA          NA
#> 279 2020-01-02 15:20:00 January 2020           NA           NA          NA
#> 280 2020-01-02 15:30:00 January 2020           NA  -69.2931578 -101.796215
#> 281 2020-01-02 15:40:00 January 2020  -83.9871453  -16.0064178 -114.455555
#> 282 2020-01-02 15:50:00 January 2020           NA           NA          NA
#> 283 2020-01-02 16:00:00 January 2020           NA           NA          NA
#> 284 2020-01-02 16:10:00 January 2020  -52.8266453   37.7687422  -87.489195
#> 285 2020-01-02 16:20:00 January 2020           NA  -95.5651678  -45.825432
#> 286 2020-01-02 16:30:00 January 2020           NA           NA          NA
#> 287 2020-01-02 16:40:00 January 2020           NA           NA          NA
#> 288 2020-01-02 16:50:00 January 2020           NA           NA          NA
#> 289 2020-01-02 17:00:00 January 2020           NA           NA          NA
#> 290 2020-01-02 17:10:00 January 2020           NA           NA          NA
#> 291 2020-01-02 17:20:00 January 2020           NA           NA          NA
#> 292 2020-01-02 17:30:00 January 2020           NA -123.1651428  -98.562765
#> 293 2020-01-02 17:40:00 January 2020           NA           NA          NA
#> 294 2020-01-02 17:50:00 January 2020           NA           NA          NA
#> 295 2020-01-02 18:00:00 January 2020 -100.2988453 -106.8486428 -104.966140
#> 296 2020-01-02 18:10:00 January 2020  -80.7748453  -18.9452928 -124.060665
#> 297 2020-01-02 18:20:00 January 2020           NA           NA          NA
#> 298 2020-01-02 18:30:00 January 2020           NA           NA          NA
#> 299 2020-01-02 18:40:00 January 2020           NA           NA          NA
#> 300 2020-01-02 18:50:00 January 2020           NA  -77.0416678  -46.092615
#> 301 2020-01-02 19:00:00 January 2020           NA           NA          NA
#> 302 2020-01-02 19:10:00 January 2020           NA           NA          NA
#> 303 2020-01-02 19:20:00 January 2020  -70.2923620           NA          NA
#> 304 2020-01-02 19:30:00 January 2020           NA           NA          NA
#> 305 2020-01-02 19:40:00 January 2020  -25.5869953           NA          NA
#> 306 2020-01-02 19:50:00 January 2020           NA           NA          NA
#> 307 2020-01-02 20:00:00 January 2020 -100.7223703  -48.1276845 -118.613048
#> 308 2020-01-02 20:10:00 January 2020           NA           NA          NA
#> 309 2020-01-02 20:20:00 January 2020           NA           NA          NA
#> 310 2020-01-02 20:30:00 January 2020           NA           NA          NA
#> 311 2020-01-02 20:40:00 January 2020           NA  -93.3103511 -101.695515
#> 312 2020-01-02 20:50:00 January 2020 -100.8773953  -45.0852178 -100.945448
#> 313 2020-01-02 21:00:00 January 2020  -96.2232753  -50.5570428  -98.775840
#> 314 2020-01-02 21:10:00 January 2020           NA           NA          NA
#> 315 2020-01-02 21:20:00 January 2020  -93.4107553  -54.0600378  -97.542675
#> 316 2020-01-02 21:30:00 January 2020           NA -112.5063178  -75.492415
#> 317 2020-01-02 21:40:00 January 2020  -78.5394153  -49.0647035  -99.966129
#> 318 2020-01-02 21:50:00 January 2020           NA           NA          NA
#> 319 2020-01-02 22:00:00 January 2020           NA           NA          NA
#> 320 2020-01-02 22:10:00 January 2020           NA           NA          NA
#> 321 2020-01-02 22:20:00 January 2020           NA           NA          NA
#> 322 2020-01-02 22:30:00 January 2020           NA           NA          NA
#> 323 2020-01-02 22:40:00 January 2020           NA           NA          NA
#> 324 2020-01-02 22:50:00 January 2020           NA           NA          NA
#> 325 2020-01-02 23:00:00 January 2020           NA           NA          NA
#> 326 2020-01-02 23:10:00 January 2020  -68.1154620           NA          NA
#> 327 2020-01-02 23:20:00 January 2020           NA           NA          NA
#> 328 2020-01-02 23:30:00 January 2020           NA           NA          NA
#> 329 2020-01-02 23:40:00 January 2020           NA           NA          NA
#> 330 2020-01-02 23:50:00 January 2020           NA           NA          NA
#> 331 2020-01-03 00:00:00 January 2020           NA           NA          NA
#> 332 2020-01-03 00:10:00 January 2020           NA           NA          NA
#> 333 2020-01-03 00:20:00 January 2020           NA           NA          NA
#> 334 2020-01-03 00:30:00 January 2020           NA           NA          NA
#> 335 2020-01-03 00:40:00 January 2020           NA           NA          NA
#> 336 2020-01-03 00:50:00 January 2020           NA           NA          NA
#> 337 2020-01-03 01:00:00 January 2020           NA           NA          NA
#> 338 2020-01-03 01:10:00 January 2020  -78.0075953  -67.4552178  -92.417615
#> 339 2020-01-03 01:20:00 January 2020           NA           NA          NA
#> 340 2020-01-03 01:30:00 January 2020           NA           NA          NA
#> 341 2020-01-03 01:40:00 January 2020           NA           NA          NA
#> 342 2020-01-03 01:50:00 January 2020           NA           NA          NA
#> 343 2020-01-03 02:00:00 January 2020           NA           NA          NA
#> 344 2020-01-03 02:10:00 January 2020  -85.0092453  -33.5585178 -105.052865
#> 345 2020-01-03 02:20:00 January 2020  -82.6753620  -44.8574178 -100.841115
#> 346 2020-01-03 02:30:00 January 2020           NA           NA          NA
#> 347 2020-01-03 02:40:00 January 2020           NA           NA          NA
#> 348 2020-01-03 02:50:00 January 2020           NA           NA          NA
#> 349 2020-01-03 03:00:00 January 2020           NA           NA          NA
#> 350 2020-01-03 03:10:00 January 2020  -68.4475287           NA          NA
#> 351 2020-01-03 03:20:00 January 2020           NA           NA          NA
#> 352 2020-01-03 03:30:00 January 2020  -81.5926203  -67.4552178  -92.417615
#> 353 2020-01-03 03:40:00 January 2020           NA           NA          NA
#> 354 2020-01-03 03:50:00 January 2020           NA           NA          NA
#> 355 2020-01-03 04:00:00 January 2020           NA           NA          NA
#> 356 2020-01-03 04:10:00 January 2020  -81.5926203  -67.4552178  -92.417615
#> 357 2020-01-03 04:20:00 January 2020           NA           NA          NA
#> 358 2020-01-03 04:30:00 January 2020           NA           NA          NA
#> 359 2020-01-03 04:40:00 January 2020           NA           NA          NA
#> 360 2020-01-03 04:50:00 January 2020           NA -112.4020178  -74.218348
#> 361 2020-01-03 05:00:00 January 2020           NA           NA          NA
#> 362 2020-01-03 05:10:00 January 2020           NA           NA          NA
#> 363 2020-01-03 05:20:00 January 2020           NA           NA          NA
#> 364 2020-01-03 05:30:00 January 2020           NA           NA          NA
#> 365 2020-01-03 05:40:00 January 2020           NA           NA          NA
#> 366 2020-01-03 05:50:00 January 2020           NA           NA          NA
#> 367 2020-01-03 06:00:00 January 2020           NA           NA          NA
#> 368 2020-01-03 06:10:00 January 2020           NA           NA          NA
#> 369 2020-01-03 06:20:00 January 2020           NA           NA          NA
#> 370 2020-01-03 06:30:00 January 2020  -78.0075953           NA          NA
#> 371 2020-01-03 06:40:00 January 2020           NA           NA          NA
#> 372 2020-01-03 06:50:00 January 2020  -78.0075953           NA          NA
#> 373 2020-01-03 07:00:00 January 2020           NA  -67.4552178  -92.417615
#> 374 2020-01-03 07:10:00 January 2020           NA           NA          NA
#> 375 2020-01-03 07:20:00 January 2020           NA           NA          NA
#> 376 2020-01-03 07:30:00 January 2020  -55.9796620    4.3169822  -94.187265
#> 377 2020-01-03 07:40:00 January 2020           NA           NA          NA
#> 378 2020-01-03 07:50:00 January 2020           NA           NA          NA
#> 379 2020-01-03 08:00:00 January 2020           NA           NA          NA
#> 380 2020-01-03 08:10:00 January 2020           NA           NA          NA
#> 381 2020-01-03 08:20:00 January 2020           NA           NA          NA
#> 382 2020-01-03 08:30:00 January 2020           NA           NA          NA
#> 383 2020-01-03 08:40:00 January 2020           NA           NA          NA
#> 384 2020-01-03 08:50:00 January 2020           NA           NA          NA
#> 385 2020-01-03 09:00:00 January 2020           NA           NA          NA
#> 386 2020-01-03 09:10:00 January 2020           NA           NA          NA
#> 387 2020-01-03 09:20:00 January 2020           NA           NA          NA
#> 388 2020-01-03 09:30:00 January 2020           NA           NA          NA
#> 389 2020-01-03 09:40:00 January 2020           NA           NA          NA
#> 390 2020-01-03 09:50:00 January 2020           NA           NA          NA
#> 391 2020-01-03 10:00:00 January 2020           NA           NA          NA
#> 392 2020-01-03 10:10:00 January 2020           NA           NA          NA
#> 393 2020-01-03 10:20:00 January 2020           NA           NA          NA
#> 394 2020-01-03 10:30:00 January 2020           NA           NA          NA
#> 395 2020-01-03 10:40:00 January 2020           NA           NA          NA
#> 396 2020-01-03 10:50:00 January 2020           NA           NA          NA
#> 397 2020-01-03 11:00:00 January 2020           NA           NA          NA
#> 398 2020-01-03 11:10:00 January 2020           NA  -67.4552178  -92.417615
#> 399 2020-01-03 11:20:00 January 2020           NA           NA          NA
#> 400 2020-01-03 11:30:00 January 2020           NA           NA          NA
#> 401 2020-01-03 11:40:00 January 2020           NA  -67.8579678  -94.606265
#> 402 2020-01-03 11:50:00 January 2020           NA           NA          NA
#> 403 2020-01-03 12:00:00 January 2020  -78.0075953  135.8546822  -59.462982
#> 404 2020-01-03 12:10:00 January 2020 -134.3301953  -29.9582928    2.675435
#> 405 2020-01-03 12:20:00 January 2020 -109.9422620  -42.5528778  -99.467335
#> 406 2020-01-03 12:30:00 January 2020  -78.1283953           NA          NA
#> 407 2020-01-03 12:40:00 January 2020  -66.6562353  -44.7841011  -97.886132
#> 408 2020-01-03 12:50:00 January 2020           NA -161.5008321  -60.488886
#> 409 2020-01-03 13:00:00 January 2020 -131.8135453           NA          NA
#> 410 2020-01-03 13:10:00 January 2020 -127.5035239 -170.9787607 -111.995486
#> 411 2020-01-03 13:20:00 January 2020           NA           NA          NA
#> 412 2020-01-03 13:30:00 January 2020 -153.4867828  -88.0685053  122.250010
#> 413 2020-01-03 13:40:00 January 2020           NA           NA          NA
#> 414 2020-01-03 13:50:00 January 2020           NA -166.5710053 -134.711977
#> 415 2020-01-03 14:00:00 January 2020           NA -130.9657303 -142.202040
#> 416 2020-01-03 14:10:00 January 2020           NA -120.9750303 -173.016540
#> 417 2020-01-03 14:20:00 January 2020           NA  -62.2079053   52.577835
#> 418 2020-01-03 14:30:00 January 2020  -31.2859753  109.7125266   40.950785
#> 419 2020-01-03 14:40:00 January 2020 -102.5978153  -46.4728956 -106.091515
#> 420 2020-01-03 14:50:00 January 2020  -12.7950287    7.3886489 -178.578715
#> 421 2020-01-03 15:00:00 January 2020           NA -122.4342678 -165.272615
#> 422 2020-01-03 15:10:00 January 2020 -119.4649753  -57.0991678  -24.389645
#> 423 2020-01-03 15:20:00 January 2020  -64.0592453    2.9279185 -134.621279
#> 424 2020-01-03 15:30:00 January 2020  -77.5311120   -0.5086178  -82.100915
#> 425 2020-01-03 15:40:00 January 2020  -33.3985525   -9.6835845 -123.632223
#> 426 2020-01-03 15:50:00 January 2020 -129.4221203  -82.7667845  -49.204107
#> 427 2020-01-03 16:00:00 January 2020  -72.9001509  -29.3838178 -197.629923
#> 428 2020-01-03 16:10:00 January 2020 -114.7773253 -112.1311261 -221.273615
#> 429 2020-01-03 16:20:00 January 2020           NA -145.2631928   -5.331898
#> 430 2020-01-03 16:30:00 January 2020           NA -158.7003095 -185.961548
#> 431 2020-01-03 16:40:00 January 2020   14.7940047 -152.5111178 -177.453248
#> 432 2020-01-03 16:50:00 January 2020 -100.0010842 -133.5011178 -132.365782
#> 433 2020-01-03 17:00:00 January 2020           NA           NA          NA
#> 434 2020-01-03 17:10:00 January 2020 -130.3788620  -69.4279178   -9.981197
#> 435 2020-01-03 17:20:00 January 2020 -120.1000287  -64.1732724  -64.432651
#> 436 2020-01-03 17:30:00 January 2020  -71.7585953    9.7277185  -80.390197
#> 437 2020-01-03 17:40:00 January 2020  -19.5280953    7.2036276 -122.926379
#> 438 2020-01-03 17:50:00 January 2020  -81.3184176    6.1184822  -18.144706
#> 439 2020-01-03 18:00:00 January 2020   35.9568713  -71.0965178  -53.870451
#> 440 2020-01-03 18:10:00 January 2020 -138.7764842  -62.8895996  -62.481297
#> 441 2020-01-03 18:20:00 January 2020   -0.1014253  137.5147913 -108.666660
#> 442 2020-01-03 18:30:00 January 2020           NA -154.4978996 -128.175479
#> 443 2020-01-03 18:40:00 January 2020           NA -163.5339269 -182.619869
#> 444 2020-01-03 18:50:00 January 2020  -46.7480287   43.3674367 -103.834379
#> 445 2020-01-03 19:00:00 January 2020  -95.9221253  -36.7270345  -80.382482
#> 446 2020-01-03 19:10:00 January 2020 -135.4592853 -115.4607261 -178.113690
#> 447 2020-01-03 19:20:00 January 2020 -121.2518153           NA          NA
#> 448 2020-01-03 19:30:00 January 2020  -64.4847653 -160.1613633  -64.989879
#> 449 2020-01-03 19:40:00 January 2020           NA -135.9589087    5.237176
#> 450 2020-01-03 19:50:00 January 2020 -120.0909509           NA          NA
#> 451 2020-01-03 20:00:00 January 2020           NA -147.7727678 -118.778195
#> 452 2020-01-03 20:10:00 January 2020  -48.0753328  -15.9281578  -71.702395
#> 453 2020-01-03 20:20:00 January 2020  -50.7473328 -125.0520878 -109.293025
#> 454 2020-01-03 20:30:00 January 2020 -108.4909176 -125.2846778 -110.694915
#> 455 2020-01-03 20:40:00 January 2020           NA -133.7671278 -161.803885
#> 456 2020-01-03 20:50:00 January 2020 -123.5262731  -44.0867478  -51.053785
#> 457 2020-01-03 21:00:00 January 2020  -40.0224398 -103.2736178 -152.509675
#> 458 2020-01-03 21:10:00 January 2020  -51.9389620  -95.9693378 -151.373075
#> 459 2020-01-03 21:20:00 January 2020  -84.4507620   37.8037640 -138.369215
#> 460 2020-01-03 21:30:00 January 2020  -99.8562509  -29.5355996  162.462494
#> 461 2020-01-03 21:40:00 January 2020 -104.1756853 -109.3488087 -148.525051
#> 462 2020-01-03 21:50:00 January 2020  -72.7226053  -44.0535345 -125.201040
#> 463 2020-01-03 22:00:00 January 2020  -88.3986044   25.3550239  -19.013282
#> 464 2020-01-03 22:10:00 January 2020   46.0174228  242.3240239   33.420135
#> 465 2020-01-03 22:20:00 January 2020  -40.8848499   60.5266322 -132.357640
#> 466 2020-01-03 22:30:00 January 2020           NA -147.1442178  -77.001323
#> 467 2020-01-03 22:40:00 January 2020 -116.3027590 -155.0765595 -124.796048
#> 468 2020-01-03 22:50:00 January 2020  -89.1433590  -95.1626845 -155.085048
#> 469 2020-01-03 23:00:00 January 2020           NA -155.6016511 -122.806273
#> 470 2020-01-03 23:10:00 January 2020           NA -155.6520178 -123.109732
#> 471 2020-01-03 23:20:00 January 2020  -88.6515287  -87.1256428  -75.270332
#> 472 2020-01-03 23:30:00 January 2020  -90.6236287  -87.8251428 -106.536140
#> 473 2020-01-03 23:40:00 January 2020           NA -135.3021761  -29.485190
#> 474 2020-01-03 23:50:00 January 2020 -147.5030078  -84.2366345  -26.567048
#> 475 2020-01-04 00:00:00 January 2020 -132.0474328  -82.6696178 -150.192698
#> 476 2020-01-04 00:10:00 January 2020           NA           NA          NA
#> 477 2020-01-04 00:20:00 January 2020           NA -106.5904178 -166.647360
#> 478 2020-01-04 00:30:00 January 2020  -62.6652096  -38.8333724  -42.388333
#> 479 2020-01-04 00:40:00 January 2020           NA -109.0881724 -153.297215
#> 480 2020-01-04 00:50:00 January 2020           NA -103.3690178 -154.877697
#> 481 2020-01-04 01:00:00 January 2020  -62.9661787  -37.8200815  -44.713860
#> 482 2020-01-04 01:10:00 January 2020           NA  -85.6075451  -19.380842
#> 483 2020-01-04 01:20:00 January 2020           NA           NA          NA
#> 484 2020-01-04 01:30:00 January 2020  -93.5626153  -37.8770578 -141.905715
#> 485 2020-01-04 01:40:00 January 2020  -51.2916120   23.2571722 -144.198595
#> 486 2020-01-04 01:50:00 January 2020           NA           NA          NA
#> 487 2020-01-04 02:00:00 January 2020           NA           NA          NA
#> 488 2020-01-04 02:10:00 January 2020           NA -102.1637289  -92.430237
#> 489 2020-01-04 02:20:00 January 2020           NA -110.5043845 -142.684704
#> 490 2020-01-04 02:30:00 January 2020  -98.8795703  -42.5318734 -147.273204
#> 491 2020-01-04 02:40:00 January 2020           NA -110.3751734 -129.188282
#> 492 2020-01-04 02:50:00 January 2020           NA           NA          NA
#> 493 2020-01-04 03:00:00 January 2020  -90.0964453  -45.7641303 -107.111077
#> 494 2020-01-04 03:10:00 January 2020           NA           NA          NA
#> 495 2020-01-04 03:20:00 January 2020           NA           NA          NA
#> 496 2020-01-04 03:30:00 January 2020  -88.4617953  -50.2160035 -115.455915
#> 497 2020-01-04 03:40:00 January 2020           NA  -97.5578607  -93.340744
#> 498 2020-01-04 03:50:00 January 2020  -43.5731953   31.6504822  -90.784715
#> 499 2020-01-04 04:00:00 January 2020           NA -106.4813511  -91.537437
#> 500 2020-01-04 04:10:00 January 2020           NA -106.1814956  -89.731059